major reorganization of config and project files to manage CLI with BYOK
