+++ .gitignore
@@ -3,8 +3,8 @@
 
 ### Python ###
 # Byte-compiled / optimized / DLL files
-Grant_Guide/__pycache__/
-Grant_Guide_config/__pycache__/
+Grant_Guide/__pycache__/*
+Grant_Guide_config/__pycache__/*
 *.py[cod]
 *$py.class
